+++ CHANGES.rst
@@ -1,6 +1,7 @@
 Unreleased
 ==========
 
+- [IMPROVED] Move tox versions to python 3.6, 3.7, 3.8
 
 0.2.1 (2018-02-19)
 ==================